Fix error in and simplify make_move endpoint, fix error in points() function
The make_move endpoint had a fair amount of repetition in the if statment handling whether or not a match had been found; this has been simplified. The make_move endpoint contained an error; it was not possible to make a match and lose the game at the same time (a situation that could occur if a user had run out of guesses). This error has been corrected. The points() function in game_logic.py contained an error that caused the total_guesses variable to be calculated incorrectly. This has been fixed.